Set up cache for items in the search result
RSS items from the search result are created with no reference to the
cache instance. This causes missing of item read/unread status. The bug
is easy to reproduce:

 * load a feed with some items
 * start a new search that finds one or several items
 * read any item from the search result
 * quit search
 * run the same search again

Expected result: already read items should NOT be marked as "unread".

In this commit we deliberately assign a reference to the cache instance
for all RSS items found in the database, the same way it's done in the
`cache::internalize_rssfeed` method. A new unit test provides steps to
reproduce behavior described above, so we can be sure that the bug is
really fixed.