Allow searching in query feeds
controller::search_for_items accepts a feedurl string and passes it down to the
cache. Since query feeds are not stored in the cache this causes all query feed
searches to return no results. The solution is to pass the feed object to
search_for_items and do a linear search from the items that have been already
loaded/matched if we are searching a query feed.

Also added an assert call in cache::search_for_items to prevent bugs like this in the future.

Fixes #313