Use new item instances for query feed search
The way search works right now, a new item instance is created for every search
result. Using an already loaded item instance doesn't work correctly because
rss_item.unload() is called after exiting from the search results for each item
which clears the description/content and since the query feed and the search
result share the same item instance, the item description gets cleared off of
the query feeds item too.

The quick solution is to simply make query feed searches create a new item
instance too, but this is ugly and should probably be replaced with a better
system in the long term.